Add Market

+++ 9cWebMarket.Frontend/Models/Api/StatModel.cs
@@ -4,6 +4,6 @@ namespace NineCWebMarket.Frontend.Models.Api;
 
 public record StatModel(
     [property: JsonPropertyName("type")] int Type,
-    [property: JsonPropertyName("value")] int Value,
+    [property: JsonPropertyName("value")] decimal Value,
     [property: JsonPropertyName("additional")] bool Additional
 );
